Use window.location in G+ comment thing
It seemed to break before due to the comment widget not liking hostnames
different than the one in window.location.
As we might be viewing on a different DNS than the one set in SITEURL
I hope this will be slightly more flexible.

Although it will still mean comments will only appear from the one DNS
location.